requirements.txt
streamlit
python-docx
PyMuPDF
mail-parser
faiss-cpu
openai
python-dotenv
tqdm
requests
fastapi
uvicorn
python-multipart
pyngrok